style-infra: add a living style guide so we can see what our style changes look like
This should really live in its own app, but I’d rather be scrappy right
now. It looks bad because our actual styling has a couple of quirks we
need to get rid of (which is causing issues like not being able to
scroll naturally on the login page). I may start with a blank bootstrap
setup and re-implement individual component styles one by one to see
what’s necessary.